Add SampleTest.EntityFrameworkCore
