+++ src/SampleTest.EntityFrameworkCore/JsonColumnObjectTest.cs
