+++ src/SampleTest.EntityFrameworkCore/OwnedEntityStampTest.cs
