+++ src/SampleTest.EntityFrameworkCore/SystemCatalogViewTest.cs
