+++ src/SampleTest.EntityFrameworkCore/TheoryDataFactory.cs
@@ -1,0 +1,11 @@
+namespace SampleTest.EntityFrameworkCore;
+
+public static class TheoryDataFactory {
+	public static TheoryData<TValue> CreateFrom<TValue>(IEnumerable<TValue> values) {
+		var data = new TheoryData<TValue>();
+		foreach (var value in values) {
+			data.Add(value);
+		}
+		return data;
+	}
+}
