+++ src/SampleTest.EntityFrameworkCore/ValueConversionEnumStringTest.cs
