+++ src/SampleTest.EntityFrameworkCore/ValueConversionImmutableTest.cs
